Small changes: * separate the v3/content/batchAnnouncement e2e tests * update/regen the swagger openapi-specs for above * update the PR template to remind ourselves to do so * minor version update for babel
